feat: add theme switcher functionality to mode toggle button
