+++ docs/dev-notes.md
@@ -3,28 +3,17 @@
 - full moon
 - chance of outside recess (weather)
 - not in the mood
-- countdown until next break
-- theme switcher
-- breaks are, thanksgiving, winter, spring
 - consider using drawer for contact button
 - consider using drawer for donate button
 - use drawer component in mobile view and dialog in desktop
-- consider making logo icon bigger
 - look into how easy it is to use stripe in a small form component in drawer.
 - consider adding drizzle-kit push --force to preview branches:
-- Add countdown selection in the navbar using a dropdown
-- Move dashboard page contents to it's own module
-- Clean up clerk appearance variables
-- look into tailwind v3 to be able to use theme switcher
-- Check to see if I can get theme color data from the theme switcher components for clerk appearance
 - iOS chrome browser bounces when scrolling to bottom and reverts to top. doesn't happen on other browsers, not even brave
-- Work on theme switcher
-  - Theme switching will be entirely client-side
-  - use a server action to set user theme choice to clerk publicMetadata sessionClaims
-  - Get the data from sessionClaims for user theme preference as the default theme in the ThemeProvider
 - Write docs on how to add new themes
-- @custom-variant dark (&:is(.dark \*)) is actually needed to prevent light themes from showing dark
 - Figure out how to apply .dark class to all dark themes
+- Update toggle mode button to work with theme switcher
+- Maybe use a server action to update session with user theme preference
+- Make Clerk appearnce variables work with new theme
 
 ```
     "vercel-build": "if [ \"$VERCEL_ENV\" = \"production\" ]; then drizzle-kit push; elif [ \"$VERCEL_ENV\" = \"preview\" ]; drizzle-kit push --force; fi && next build",
